feat(api): Partners now scoped the customer

+++ src/api/partners/customers/[id]/addresses/[addressId]/route.ts
@@ -1,12 +1,17 @@
 import { AuthenticatedMedusaRequest, MedusaResponse } from "@medusajs/framework/http"
 import { Modules } from "@medusajs/framework/utils"
-import { getPartnerStore } from "../../../../helpers"
+import { validatePartnerEntityOwnership } from "../../../../helpers"
 
 export const GET = async (
   req: AuthenticatedMedusaRequest,
   res: MedusaResponse
 ) => {
-  await getPartnerStore(req.auth_context, req.scope)
+  await validatePartnerEntityOwnership(
+    req.auth_context,
+    "customers",
+    req.params.id,
+    req.scope
+  )
 
   const customerService = req.scope.resolve(Modules.CUSTOMER) as any
   const address = await customerService.retrieveCustomerAddress(req.params.addressId)
@@ -18,7 +23,12 @@ export const POST = async (
   req: AuthenticatedMedusaRequest,
   res: MedusaResponse
 ) => {
-  await getPartnerStore(req.auth_context, req.scope)
+  await validatePartnerEntityOwnership(
+    req.auth_context,
+    "customers",
+    req.params.id,
+    req.scope
+  )
 
   const customerService = req.scope.resolve(Modules.CUSTOMER) as any
   const address = await customerService.updateCustomerAddresses(
@@ -33,7 +43,12 @@ export const DELETE = async (
   req: AuthenticatedMedusaRequest,
   res: MedusaResponse
 ) => {
-  await getPartnerStore(req.auth_context, req.scope)
+  await validatePartnerEntityOwnership(
+    req.auth_context,
+    "customers",
+    req.params.id,
+    req.scope
+  )
 
   const customerService = req.scope.resolve(Modules.CUSTOMER) as any
   await customerService.deleteCustomerAddresses(req.params.addressId)
